+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * [TASK] updates dependencies in composer.json and ext_emconf.php for TYPO3 8.7
 * [TASK] updates license in composer.json for packagist compatibility
 * [TASK] updates Cache/CacheChain.php with new method for TYPO3 8.7
+* [TASK] updates Backend/Toolbar/ClearCacheMenu for TYPO3 8.7
+* [TASK] adds Backend CSS for TYPO3 8.7
 
 3.0.5 - 2017-06-21
 ------------------